161: Rename profile to profiles
